lib/pm: fix MISRA.SHIFT.RANGE.2012
The right hand operand of a shift operator shall lie in the
range zero to one less than the width in bits of the essential
type of the left hand operand.

Fixed by adding requied typecasting.

Fixes: SYSFW-6414

+++ pm/core/init.c
@@ -101,9 +101,9 @@ s32 pm_init(void)
 	} while (!done && progress);
 
 	if (!progress) {
-		pm_trace(TRACE_PM_ACTION_FAIL | TRACE_PM_ACTION_INIT, TRACE_PM_VAL_INIT_DEFER);
+		pm_trace(TRACE_PM_ACTION_FAIL | TRACE_PM_ACTION_INIT, (u32) TRACE_PM_VAL_INIT_DEFER);
 	} else if (errors) {
-		pm_trace(TRACE_PM_ACTION_FAIL | TRACE_PM_ACTION_INIT, TRACE_PM_VAL_INIT_ERROR);
+		pm_trace(TRACE_PM_ACTION_FAIL | TRACE_PM_ACTION_INIT, (u32) TRACE_PM_VAL_INIT_ERROR);
 	} else {
 		pm_trace(TRACE_PM_ACTION_INIT, 0x0U);
 	}
