+++ pm/soc/j7200/clk_mux_j7_ahclko.c
@@ -115,7 +115,7 @@ static sbool clk_mux_j7_ahclko_set_parent(struct clk *clock_ptr, u8 new_parent)
 			ret = SFALSE;
 		}
 		pm_trace(TRACE_PM_ACTION_CLOCK_SET_PARENT,
-			 ((new_parent << TRACE_PM_VAL_CLOCK_VAL_SHIFT) &
+			 (((u16) new_parent << TRACE_PM_VAL_CLOCK_VAL_SHIFT) &
 			  TRACE_PM_VAL_CLOCK_VAL_MASK) |
 			 ((clk_id(clock_ptr) << TRACE_PM_VAL_CLOCK_ID_SHIFT) &
 			  TRACE_PM_VAL_CLOCK_ID_MASK));
