+++ pm/soc/j721s2/clk_mux_j7_dpi.c
@@ -123,7 +123,7 @@ static sbool clk_mux_j7_dpi_0_pclk_set_parent(struct clk *clock_ptr	TRACE_ONLY,
 
 	if (ret) {
 		pm_trace(TRACE_PM_ACTION_CLOCK_SET_PARENT,
-			 ((new_parent << TRACE_PM_VAL_CLOCK_VAL_SHIFT) &
+			 (((u16) new_parent << TRACE_PM_VAL_CLOCK_VAL_SHIFT) &
 			  TRACE_PM_VAL_CLOCK_VAL_MASK) |
 			 ((clk_id(clock_ptr) << TRACE_PM_VAL_CLOCK_ID_SHIFT) &
 			  TRACE_PM_VAL_CLOCK_ID_MASK));
@@ -220,7 +220,7 @@ static sbool clk_mux_j7_dpi_2_pclk_set_parent(struct clk *clock_ptr	TRACE_ONLY,
 
 	if (ret) {
 		pm_trace(TRACE_PM_ACTION_CLOCK_SET_PARENT,
-			 ((new_parent << TRACE_PM_VAL_CLOCK_VAL_SHIFT) &
+			 (((u16) new_parent << TRACE_PM_VAL_CLOCK_VAL_SHIFT) &
 			  TRACE_PM_VAL_CLOCK_VAL_MASK) |
 			 ((clk_id(clock_ptr) << TRACE_PM_VAL_CLOCK_ID_SHIFT) &
 			  TRACE_PM_VAL_CLOCK_ID_MASK));
@@ -329,7 +329,7 @@ static sbool clk_mux_j7_dpi_3_pclk_set_parent(struct clk *clock_ptr, u8 new_pare
 			ret = SFALSE;
 		}
 		pm_trace(TRACE_PM_ACTION_CLOCK_SET_PARENT,
-			 ((new_parent << TRACE_PM_VAL_CLOCK_VAL_SHIFT) &
+			 (((u16) new_parent << TRACE_PM_VAL_CLOCK_VAL_SHIFT) &
 			  TRACE_PM_VAL_CLOCK_VAL_MASK) |
 			 ((clk_id(clock_ptr) << TRACE_PM_VAL_CLOCK_ID_SHIFT) &
 			  TRACE_PM_VAL_CLOCK_ID_MASK));
@@ -426,7 +426,7 @@ static sbool clk_mux_j7_dpi_1_clk_out_set_parent(struct clk *clock_ptr	TRACE_ONL
 
 	if (ret) {
 		pm_trace(TRACE_PM_ACTION_CLOCK_SET_PARENT,
-			 ((new_parent << TRACE_PM_VAL_CLOCK_VAL_SHIFT) &
+			 (((u16) new_parent << TRACE_PM_VAL_CLOCK_VAL_SHIFT) &
 			  TRACE_PM_VAL_CLOCK_VAL_MASK) |
 			 ((clk_id(clock_ptr) << TRACE_PM_VAL_CLOCK_ID_SHIFT) &
 			  TRACE_PM_VAL_CLOCK_ID_MASK));
